VM:support ssh keygen for vm/vmss create
