+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -1150,7 +1150,8 @@ def list_container_services(client, resource_group_name=None):
 def create_vm(vm_name, resource_group_name, image,
               size='Standard_DS1', location=None, tags=None, no_wait=False,
               authentication_type=None, admin_password=None, admin_username=getpass.getuser(),
-              ssh_dest_key_path=None, ssh_key_value=None, availability_set=None,
+              ssh_dest_key_path=None, ssh_key_value=None, generate_ssh_keys=False,
+              availability_set=None,
               nics=None, nsg=None, nsg_rule=None,
               private_ip_address=None,
               public_ip_address=None, public_ip_address_allocation='dynamic',
@@ -1279,7 +1280,7 @@ def create_vmss(vmss_name, resource_group_name, image,
                 location=None, tags=None, upgrade_policy_mode='manual', validate=False,
                 admin_username=getpass.getuser(), admin_password=None, authentication_type=None,
                 vm_sku="Standard_D1_v2", no_wait=False,
-                ssh_dest_key_path=None, ssh_key_value=None,
+                ssh_dest_key_path=None, ssh_key_value=None, generate_ssh_keys=False,
                 load_balancer=None, backend_pool_name=None, nat_pool_name=None, backend_port=None,
                 public_ip_address=None, public_ip_address_allocation='dynamic',
                 public_ip_address_dns_name=None,
